+++ Jessica/ErrMsg.py
@@ -1,5 +1,8 @@
-class ErrMsg:
+class Core:
 
     def __init__(self):
         self.UnsupportStr = "不支持输入字符"
         self.UnableAccessThisFile = "另一个程序正在使用此文件，进程无法访问。"
+
+
+ErrMsg = Core()
