+++ Jessica/Format.py
@@ -1,5 +1,6 @@
 import time
 
+
 class CurrentTime:
     YYYYMMDDHHMMSS = time.strftime("%Y-%m-%d-%H-%M-%S", time.localtime())
     YYYYMMDD = time.strftime("%Y%m%d", time.localtime())
